pyproject.toml
[project]
name = "confluence-sotd-calendar"
version = "0.1.0"
description = "CLI to create Support of the Day calendar events in Confluence Team Calendars"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "requests>=2.31",
    "python-dotenv>=1.0",
]

[tool.ruff]
line-length = 100
